Rename `mocksContainer/` to `mocks-container`
