+++ test/test-utils/mocks/mocks-container/mock-input-text.ts
@@ -1,4 +1,4 @@
-import { MockText } from "#test/test-utils/mocks/mocksContainer/mock-text";
+import { MockText } from "#test/test-utils/mocks/mocks-container/mock-text";
 
 export class MockInputText extends MockText {
   public inputType: string;
