+++ test/test-utils/mocks/mocks-container/mock-polygon.ts
@@ -1,4 +1,4 @@
-import { MockContainer } from "#test/test-utils/mocks/mocksContainer/mock-container";
+import { MockContainer } from "#test/test-utils/mocks/mocks-container/mock-container";
 
 export class MockPolygon extends MockContainer {
   constructor(textureManager, x, y, _content, _fillColor, _fillAlpha) {
