+++ test/test-utils/mocks/mocks-container/mock-sprite.ts
